Moving NameHeader definition to separate file
