+++ DataFormats/Headers/src/DataHeader.cxx
@@ -77,10 +77,6 @@ const uint32_t o2::Header::DataHeader::sVersion = 1;
 const o2::Header::HeaderType o2::Header::DataHeader::sHeaderType = String2<uint64_t>("DataHead");
 const o2::Header::SerializationMethod o2::Header::DataHeader::sSerializationMethod = o2::Header::gSerializationMethodNone;
 
-//storage fr NameHeader static
-template <>
-const o2::Header::HeaderType o2::Header::NameHeader<0>::sHeaderType = "NameHead";
-
 using namespace o2::Header;
 
 //__________________________________________________________________________________________________
